:recycle: カードコンポーネントをリネーム
